cast back error
casted from ull to int which has half the bytes thus cannot represent some ull numbers. Fixed by intermittently casting to long long instead

+++ kLibrary/Source Files/Maths/kAlgorithms.hpp
@@ -35,7 +35,7 @@ namespace kmaths
 		}
 
 		const auto valuePlusDpsByAcc = (value + dpShifts) * accuracy;
-		const T roundedValue = CAST(T, CAST(int, valuePlusDpsByAcc)) / accuracy;
+		const T roundedValue = CAST(T, CAST(long long, valuePlusDpsByAcc)) / accuracy;
 		return roundedValue;
 	}
